more deno and supabase fn cleanup
