+++ .source/index.ts
@@ -1,236 +1,25 @@
 // @ts-nocheck -- skip type checking
-import * as docs_20 from '../content/docs/studio/sync-events.mdx?collection=docs&hash=1753296249963';
-import * as docs_19 from '../content/docs/studio/settings.mdx?collection=docs&hash=1753296249963';
-import * as docs_18 from '../content/docs/studio/prompts.mdx?collection=docs&hash=1753296249963';
-import * as docs_17 from '../content/docs/studio/project.mdx?collection=docs&hash=1753296249963';
-import * as docs_16 from '../content/docs/studio/organization.mdx?collection=docs&hash=1753296249963';
-import * as docs_15 from '../content/docs/studio/metrics.mdx?collection=docs&hash=1753296249963';
-import * as docs_14 from '../content/docs/studio/logs.mdx?collection=docs&hash=1753296249963';
-import * as docs_13 from '../content/docs/studio/globals.mdx?collection=docs&hash=1753296249963';
-import * as docs_12 from '../content/docs/studio/alerts.mdx?collection=docs&hash=1753296249963';
-import * as docs_11 from '../content/docs/studio/account.mdx?collection=docs&hash=1753296249963';
-import * as docs_10 from '../content/docs/sdk/reference.mdx?collection=docs&hash=1753296249963';
-import * as docs_9 from '../content/docs/sdk/index.mdx?collection=docs&hash=1753296249963';
-import * as docs_8 from '../content/docs/sdk/basic-usage.mdx?collection=docs&hash=1753296249963';
-import * as docs_7 from '../content/docs/sdk/agentsmith-folder.mdx?collection=docs&hash=1753296249963';
-import * as docs_6 from '../content/docs/sdk/advanced-usage.mdx?collection=docs&hash=1753296249963';
-import * as docs_5 from '../content/docs/core-concepts/self-hosting.mdx?collection=docs&hash=1753296249963';
-import * as docs_4 from '../content/docs/core-concepts/architecture.mdx?collection=docs&hash=1753296249963';
-import * as docs_3 from '../content/docs/syncing.mdx?collection=docs&hash=1753296249963';
-import * as docs_2 from '../content/docs/quick-start.mdx?collection=docs&hash=1753296249963';
-import * as docs_1 from '../content/docs/prompt-authoring.mdx?collection=docs&hash=1753296249963';
-import * as docs_0 from '../content/docs/index.mdx?collection=docs&hash=1753296249963';
-import { _runtime } from 'fumadocs-mdx';
-import * as _source from '../source.config';
-export const docs = _runtime.docs<typeof _source.docs>(
-  [
-    {
-      info: {
-        path: 'index.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/index.mdx',
-      },
-      data: docs_0,
-    },
-    {
-      info: {
-        path: 'prompt-authoring.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/prompt-authoring.mdx',
-      },
-      data: docs_1,
-    },
-    {
-      info: {
-        path: 'quick-start.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/quick-start.mdx',
-      },
-      data: docs_2,
-    },
-    {
-      info: {
-        path: 'syncing.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/syncing.mdx',
-      },
-      data: docs_3,
-    },
-    {
-      info: {
-        path: 'core-concepts/architecture.mdx',
-        absolutePath:
-          '/Users/lanzo/projects/agentsmith/content/docs/core-concepts/architecture.mdx',
-      },
-      data: docs_4,
-    },
-    {
-      info: {
-        path: 'core-concepts/self-hosting.mdx',
-        absolutePath:
-          '/Users/lanzo/projects/agentsmith/content/docs/core-concepts/self-hosting.mdx',
-      },
-      data: docs_5,
-    },
-    {
-      info: {
-        path: 'sdk/advanced-usage.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/sdk/advanced-usage.mdx',
-      },
-      data: docs_6,
-    },
-    {
-      info: {
-        path: 'sdk/agentsmith-folder.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/sdk/agentsmith-folder.mdx',
-      },
-      data: docs_7,
-    },
-    {
-      info: {
-        path: 'sdk/basic-usage.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/sdk/basic-usage.mdx',
-      },
-      data: docs_8,
-    },
-    {
-      info: {
-        path: 'sdk/index.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/sdk/index.mdx',
-      },
-      data: docs_9,
-    },
-    {
-      info: {
-        path: 'sdk/reference.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/sdk/reference.mdx',
-      },
-      data: docs_10,
-    },
-    {
-      info: {
-        path: 'studio/account.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/account.mdx',
-      },
-      data: docs_11,
-    },
-    {
-      info: {
-        path: 'studio/alerts.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/alerts.mdx',
-      },
-      data: docs_12,
-    },
-    {
-      info: {
-        path: 'studio/globals.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/globals.mdx',
-      },
-      data: docs_13,
-    },
-    {
-      info: {
-        path: 'studio/logs.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/logs.mdx',
-      },
-      data: docs_14,
-    },
-    {
-      info: {
-        path: 'studio/metrics.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/metrics.mdx',
-      },
-      data: docs_15,
-    },
-    {
-      info: {
-        path: 'studio/organization.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/organization.mdx',
-      },
-      data: docs_16,
-    },
-    {
-      info: {
-        path: 'studio/project.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/project.mdx',
-      },
-      data: docs_17,
-    },
-    {
-      info: {
-        path: 'studio/prompts.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/prompts.mdx',
-      },
-      data: docs_18,
-    },
-    {
-      info: {
-        path: 'studio/settings.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/settings.mdx',
-      },
-      data: docs_19,
-    },
-    {
-      info: {
-        path: 'studio/sync-events.mdx',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/sync-events.mdx',
-      },
-      data: docs_20,
-    },
-  ],
-  [
-    {
-      info: {
-        path: 'meta.json',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/meta.json',
-      },
-      data: {
-        title: 'Documentation',
-        pages: [
-          'index',
-          'quick-start',
-          '---',
-          'prompt-authoring',
-          'syncing',
-          'core-concepts',
-          'studio',
-          'sdk',
-        ],
-      },
-    },
-    {
-      info: {
-        path: 'core-concepts/meta.json',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/core-concepts/meta.json',
-      },
-      data: { title: 'Core Concepts', pages: ['architecture', 'self-hosting'] },
-    },
-    {
-      info: {
-        path: 'sdk/meta.json',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/sdk/meta.json',
-      },
-      data: {
-        title: 'SDK',
-        pages: ['index', 'basic-usage', 'advanced-usage', 'agentsmith-folder', '---', 'reference'],
-      },
-    },
-    {
-      info: {
-        path: 'studio/meta.json',
-        absolutePath: '/Users/lanzo/projects/agentsmith/content/docs/studio/meta.json',
-      },
-      data: {
-        title: 'Studio',
-        pages: [
-          'project',
-          'prompts',
-          'logs',
-          'metrics',
-          'sync-events',
-          'globals',
-          'account',
-          'organization',
-          'alerts',
-          'settings',
-        ],
-      },
-    },
-  ],
-);
+import * as docs_20 from "../content/docs/studio/sync-events.mdx?collection=docs&hash=1753296249963"
+import * as docs_19 from "../content/docs/studio/settings.mdx?collection=docs&hash=1753296249963"
+import * as docs_18 from "../content/docs/studio/prompts.mdx?collection=docs&hash=1753296249963"
+import * as docs_17 from "../content/docs/studio/project.mdx?collection=docs&hash=1753296249963"
+import * as docs_16 from "../content/docs/studio/organization.mdx?collection=docs&hash=1753296249963"
+import * as docs_15 from "../content/docs/studio/metrics.mdx?collection=docs&hash=1753296249963"
+import * as docs_14 from "../content/docs/studio/logs.mdx?collection=docs&hash=1753296249963"
+import * as docs_13 from "../content/docs/studio/globals.mdx?collection=docs&hash=1753296249963"
+import * as docs_12 from "../content/docs/studio/alerts.mdx?collection=docs&hash=1753296249963"
+import * as docs_11 from "../content/docs/studio/account.mdx?collection=docs&hash=1753296249963"
+import * as docs_10 from "../content/docs/sdk/reference.mdx?collection=docs&hash=1753296249963"
+import * as docs_9 from "../content/docs/sdk/index.mdx?collection=docs&hash=1753296249963"
+import * as docs_8 from "../content/docs/sdk/basic-usage.mdx?collection=docs&hash=1753296249963"
+import * as docs_7 from "../content/docs/sdk/agentsmith-folder.mdx?collection=docs&hash=1753296249963"
+import * as docs_6 from "../content/docs/sdk/advanced-usage.mdx?collection=docs&hash=1753296249963"
+import * as docs_5 from "../content/docs/core-concepts/self-hosting.mdx?collection=docs&hash=1753296249963"
+import * as docs_4 from "../content/docs/core-concepts/architecture.mdx?collection=docs&hash=1753296249963"
+import * as docs_3 from "../content/docs/syncing.mdx?collection=docs&hash=1753296249963"
+import * as docs_2 from "../content/docs/quick-start.mdx?collection=docs&hash=1753296249963"
+import * as docs_1 from "../content/docs/prompt-authoring.mdx?collection=docs&hash=1753296249963"
+import * as docs_0 from "../content/docs/index.mdx?collection=docs&hash=1753296249963"
+import { _runtime } from "fumadocs-mdx"
+import * as _source from "../source.config"
+export const docs = _runtime.docs<typeof _source.docs>([{ info: {"path":"index.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/index.mdx"}, data: docs_0 }, { info: {"path":"prompt-authoring.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/prompt-authoring.mdx"}, data: docs_1 }, { info: {"path":"quick-start.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/quick-start.mdx"}, data: docs_2 }, { info: {"path":"syncing.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/syncing.mdx"}, data: docs_3 }, { info: {"path":"core-concepts/architecture.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/core-concepts/architecture.mdx"}, data: docs_4 }, { info: {"path":"core-concepts/self-hosting.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/core-concepts/self-hosting.mdx"}, data: docs_5 }, { info: {"path":"sdk/advanced-usage.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/sdk/advanced-usage.mdx"}, data: docs_6 }, { info: {"path":"sdk/agentsmith-folder.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/sdk/agentsmith-folder.mdx"}, data: docs_7 }, { info: {"path":"sdk/basic-usage.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/sdk/basic-usage.mdx"}, data: docs_8 }, { info: {"path":"sdk/index.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/sdk/index.mdx"}, data: docs_9 }, { info: {"path":"sdk/reference.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/sdk/reference.mdx"}, data: docs_10 }, { info: {"path":"studio/account.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/account.mdx"}, data: docs_11 }, { info: {"path":"studio/alerts.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/alerts.mdx"}, data: docs_12 }, { info: {"path":"studio/globals.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/globals.mdx"}, data: docs_13 }, { info: {"path":"studio/logs.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/logs.mdx"}, data: docs_14 }, { info: {"path":"studio/metrics.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/metrics.mdx"}, data: docs_15 }, { info: {"path":"studio/organization.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/organization.mdx"}, data: docs_16 }, { info: {"path":"studio/project.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/project.mdx"}, data: docs_17 }, { info: {"path":"studio/prompts.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/prompts.mdx"}, data: docs_18 }, { info: {"path":"studio/settings.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/settings.mdx"}, data: docs_19 }, { info: {"path":"studio/sync-events.mdx","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/sync-events.mdx"}, data: docs_20 }], [{"info":{"path":"meta.json","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/meta.json"},"data":{"title":"Documentation","pages":["index","quick-start","---","prompt-authoring","syncing","core-concepts","studio","sdk"]}}, {"info":{"path":"core-concepts/meta.json","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/core-concepts/meta.json"},"data":{"title":"Core Concepts","pages":["architecture","self-hosting"]}}, {"info":{"path":"sdk/meta.json","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/sdk/meta.json"},"data":{"title":"SDK","pages":["index","basic-usage","advanced-usage","agentsmith-folder","---","reference"]}}, {"info":{"path":"studio/meta.json","absolutePath":"/Users/lanzo/projects/agentsmith/content/docs/studio/meta.json"},"data":{"title":"Studio","pages":["project","prompts","logs","metrics","sync-events","globals","account","organization","alerts","settings"]}}])
