+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
   "deno.enable": true,
   "deno.enablePaths": ["./supabase/functions"],
-  "deno.importMap": "./supabase/functions/import_map.json"
+  "deno.importMap": "./supabase/functions/deno.json"
 }
