--- deno.lock
@@ -1,103 +1,0 @@
-{
-  "version": "5",
-  "specifiers": {
-    "jsr:@supabase/functions-js@*": "2.4.5"
-  },
-  "jsr": {
-    "@supabase/functions-js@2.4.5": {
-      "integrity": "ead2df253fe0c3b0b682c5f25076978ab62c10afff4d04222451a90ba512c394"
-    }
-  },
-  "workspace": {
-    "packageJson": {
-      "dependencies": [
-        "npm:@hookform/resolvers@^4.1.3",
-        "npm:@next/bundle-analyzer@^15.3.3",
-        "npm:@next/third-parties@^15.1.6",
-        "npm:@octokit/webhooks@^13.7.4",
-        "npm:@radix-ui/react-accordion@^1.2.3",
-        "npm:@radix-ui/react-alert-dialog@^1.1.14",
-        "npm:@radix-ui/react-aspect-ratio@^1.1.7",
-        "npm:@radix-ui/react-avatar@^1.1.3",
-        "npm:@radix-ui/react-checkbox@^1.1.4",
-        "npm:@radix-ui/react-collapsible@^1.1.3",
-        "npm:@radix-ui/react-dialog@^1.1.6",
-        "npm:@radix-ui/react-dropdown-menu@^2.1.6",
-        "npm:@radix-ui/react-label@^2.1.2",
-        "npm:@radix-ui/react-navigation-menu@^1.2.5",
-        "npm:@radix-ui/react-popover@^1.1.14",
-        "npm:@radix-ui/react-progress@^1.1.7",
-        "npm:@radix-ui/react-scroll-area@^1.2.3",
-        "npm:@radix-ui/react-select@^2.1.6",
-        "npm:@radix-ui/react-separator@^1.1.2",
-        "npm:@radix-ui/react-slot@^1.2.3",
-        "npm:@radix-ui/react-switch@^1.1.3",
-        "npm:@radix-ui/react-tabs@^1.1.3",
-        "npm:@radix-ui/react-tooltip@^1.1.8",
-        "npm:@supabase/ssr@latest",
-        "npm:@supabase/stripe-sync-engine@0.39",
-        "npm:@supabase/supabase-js@latest",
-        "npm:@tailwindcss/postcss@^4.0.9",
-        "npm:@types/jest@^29.5.14",
-        "npm:@types/lodash.isequal@^4.5.8",
-        "npm:@types/lodash.merge@^4.6.9",
-        "npm:@types/mdx@^2.0.13",
-        "npm:@types/node@22.10.2",
-        "npm:@types/nunjucks@^3.2.6",
-        "npm:@types/prismjs@^1.26.5",
-        "npm:@types/react-dom@19.0.2",
-        "npm:@types/react@^19.0.2",
-        "npm:@types/vivus@~0.4.7",
-        "npm:class-variance-authority@0.7",
-        "npm:clsx@^2.1.1",
-        "npm:date-fns@^4.1.0",
-        "npm:dotenv@^16.4.7",
-        "npm:eventsource-client@^1.1.3",
-        "npm:eventsource-parser@^3.0.2",
-        "npm:fast-jwt@^5.0.5",
-        "npm:fumadocs-core@^15.6.1",
-        "npm:fumadocs-mdx@^11.6.10",
-        "npm:fumadocs-typescript@^4.0.6",
-        "npm:fumadocs-ui@^15.6.1",
-        "npm:input-otp@^1.4.2",
-        "npm:jest@^29.7.0",
-        "npm:js-tiktoken@^1.0.20",
-        "npm:lodash.isequal@^4.5.0",
-        "npm:lodash.merge@^4.6.2",
-        "npm:lucide-react@0.468",
-        "npm:mermaid@^11.9.0",
-        "npm:next-themes@~0.4.6",
-        "npm:next@^15.1.6",
-        "npm:nunjucks@^3.2.4",
-        "npm:octokit@^4.1.2",
-        "npm:pino@^9.7.0",
-        "npm:postcss@8.4.49",
-        "npm:posthog-js@^1.215.1",
-        "npm:prettier@^3.3.3",
-        "npm:prismjs@^1.29.0",
-        "npm:react-day-picker@^9.8.0",
-        "npm:react-diff-viewer-continued-react19@1",
-        "npm:react-dom@19.0.0",
-        "npm:react-hook-form@^7.54.2",
-        "npm:react-markdown@^10.1.0",
-        "npm:react-simple-code-editor@~0.14.1",
-        "npm:react@19.0.0",
-        "npm:recharts@^3.0.2",
-        "npm:smee-client@^3.1.1",
-        "npm:sonner@^2.0.1",
-        "npm:stripe@^18.3.0",
-        "npm:supabase@^2.9.6",
-        "npm:tailwind-merge@^2.5.2",
-        "npm:tailwindcss-animate@^1.0.7",
-        "npm:tailwindcss@^4.0.9",
-        "npm:ts-jest@^29.2.5",
-        "npm:ts-morph@^25.0.1",
-        "npm:tsd@0.32",
-        "npm:tw-animate-css@^1.3.0",
-        "npm:typescript@5.7.2",
-        "npm:vivus@~0.4.6",
-        "npm:zod@^3.24.2"
-      ]
-    }
-  }
-}
