--- supabase/functions/import_map.json
@@ -1,3 +1,0 @@
-{
-  "imports": {}
-}
