+++ tsconfig.json
@@ -27,5 +27,5 @@
     "baseUrl": "./"
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-  "exclude": ["node_modules", "ts-sdk", "__tests__", "**/*.test-d.ts", "supabase/functions"]
+  "exclude": ["node_modules", "ts-sdk", "__tests__", "**/*.test-d.ts", "supabase/functions/**/*.ts"]
 }
